moved all startup util logic out of DataStructureUtil and into StartupUtil
